WIP: start to split operator/__init__.py into multiple pieces
